crawler/python/requirements.txt
scrapy
scrapy-playwright
sqlalchemy
mysql-connector-python
python-dotenv